Dev testing changes

+++ src/main/java/com/dpw/runner/shipment/services/dto/v1/response/CoLoadingMAWBDetailsResponse.java
@@ -10,7 +10,7 @@ public class CoLoadingMAWBDetailsResponse implements Serializable {
     @JsonProperty("Id")
     private Long id;
     @JsonProperty("ParentTenantId")
-    private Long parentTenantId;
+    private Integer parentTenantId;
     @JsonProperty("ChildTenantId")
-    private Long childTenantId;
+    private Integer childTenantId;
 }
